+++ examples/with_inline_policy/data.tf
@@ -1,5 +1,7 @@
 data "aws_iam_policy_document" "ec2_assume_role_policy" {
+
   statement {
+    sid     = "ExampleSTSAssumeRole"
     actions = ["sts:AssumeRole"]
 
     principals {
@@ -11,6 +13,7 @@ data "aws_iam_policy_document" "ec2_assume_role_policy" {
 
 data "aws_iam_policy_document" "inline_policy" {
   statement {
+    sid       = "ExampleInlineEC2DescribePolicy"
     actions   = ["ec2:DescribeAccountAttributes"]
     resources = ["*"]
   }